api/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf-spectacular==0.28.0
django-cors-headers==4.6.0
redis==5.2.1
psycopg[binary]==3.2.5
gunicorn==22.0.0
whitenoise==6.7.0
